move auth/components to shared/components
