+++ src/backoffice/shared/components/input-user-group/input-user-group.element.ts
@@ -1,12 +1,9 @@
 import { UUITextStyles } from '@umbraco-ui/uui-css';
 import { css, html, nothing } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
-import { UmbInputListBase } from '../../../backoffice/shared/components/input-list-base/input-list-base';
+import { UmbInputListBase } from '../input-list-base/input-list-base';
+import { UmbUserGroupStore, UMB_USER_GROUP_STORE_CONTEXT_TOKEN } from '../../../users/user-groups/user-group.store';
 import type { UserGroupEntity } from '@umbraco-cms/models';
-import {
-	UmbUserGroupStore,
-	UMB_USER_GROUP_STORE_CONTEXT_TOKEN,
-} from 'src/backoffice/users/user-groups/user-group.store';
 
 @customElement('umb-input-user-group')
 export class UmbInputPickerUserGroupElement extends UmbInputListBase {
