+++ src/backoffice/shared/components/input-user-group/input-user-group.test.ts
