+++ src/backoffice/shared/components/input-user/input-user.element.ts
@@ -1,9 +1,9 @@
 import { UUITextStyles } from '@umbraco-ui/uui-css';
 import { css, html, nothing, PropertyValueMap } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
-import { UmbInputListBase } from '../../../backoffice/shared/components/input-list-base/input-list-base';
+import { UmbInputListBase } from '../input-list-base/input-list-base';
+import { UmbUserStore, UMB_USER_STORE_CONTEXT_TOKEN } from '../../../users/users/user.store';
 import type { UserEntity } from '@umbraco-cms/models';
-import { UmbUserStore, UMB_USER_STORE_CONTEXT_TOKEN } from 'src/backoffice/users/users/user.store';
 
 @customElement('umb-input-user')
 export class UmbPickerUserElement extends UmbInputListBase {
