+++ src/backoffice/users/user-section/views/users/list-view-layouts/grid/workspace-view-users-grid.element.ts
@@ -4,8 +4,8 @@ import { customElement, state } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import type { UmbSectionViewUsersElement } from '../../section-view-users.element';
-import { getTagLookAndColor } from '../../../../../../../auth/utils';
 import { UmbUserGroupStore, UMB_USER_GROUP_STORE_CONTEXT_TOKEN } from '../../../../../user-groups/user-group.store';
+import { getTagLookAndColor } from '@umbraco-cms/utils';
 import type { UserDetails, UserEntity, UserGroupEntity } from '@umbraco-cms/models';
 import { UmbLitElement } from '@umbraco-cms/element';
