+++ src/backoffice/users/user-section/views/users/workspace-view-users-create.element.ts
@@ -2,15 +2,15 @@ import { css, html, nothing } from 'lit';
 import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, query, state } from 'lit/decorators.js';
 import { UUIInputPasswordElement } from '@umbraco-ui/uui';
-import { UmbInputPickerUserGroupElement } from 'src/auth/components/input-user-group/input-user-group.element';
+import { UmbInputPickerUserGroupElement } from '../../../../shared/components/input-user-group/input-user-group.element';
+import { UmbModalLayoutElement } from '../../../../../core/modal';
+import { UmbUserStore, UMB_USER_STORE_CONTEXT_TOKEN } from '../../../users/user.store';
 import type { UserDetails } from '@umbraco-cms/models';
 import {
 	UmbNotificationDefaultData,
 	UmbNotificationService,
 	UMB_NOTIFICATION_SERVICE_CONTEXT_TOKEN,
 } from '@umbraco-cms/notification';
-import { UmbModalLayoutElement } from 'src/core/modal';
-import { UmbUserStore, UMB_USER_STORE_CONTEXT_TOKEN } from 'src/backoffice/users/users/user.store';
 
 export type UsersViewType = 'list' | 'grid';
 @customElement('umb-workspace-view-users-create')
