+++ src/backoffice/users/users/workspace/user-workspace.element.ts
@@ -6,15 +6,15 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 import { repeat } from 'lit/directives/repeat.js';
 
 import { distinctUntilChanged } from 'rxjs';
-import { getTagLookAndColor } from '../../../../auth/utils';
 
 import { UmbCurrentUserStore, UMB_CURRENT_USER_STORE_CONTEXT_TOKEN } from '../../current-user/current-user.store';
 import type { UmbModalService } from '../../../../core/modal';
 import type { UmbWorkspaceEntityElement } from '../../../shared/components/workspace/workspace-entity-element.interface';
 import { UmbWorkspaceUserContext } from './user-workspace.context';
+import { getTagLookAndColor } from '@umbraco-cms/utils';
 import type { UserDetails } from '@umbraco-cms/models';
 
-import '../../../../auth/components/input-user-group/input-user-group.element';
+import '../../../shared/components/input-user-group/input-user-group.element';
 import '../../../shared/property-editors/uis/document-picker/property-editor-ui-document-picker.element';
 import '../../../shared/components/workspace/workspace-layout/workspace-layout.element';
 import { UmbLitElement } from '@umbraco-cms/element';
